frontend added

+++ backend/controllers/goalController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc    Get goals
 // @route   GET /api/goals 
@@ -29,7 +30,7 @@ const setGoal =  asyncHandler(async (req, res) => {
 })
 
 // @desc    Update goals
-// @route   UPDATE /api/goal/:id 
+// @route   UPDATE /api/goals/:id 
 // @access  Private
 const updateGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id)
@@ -37,6 +38,18 @@ const updateGoal = asyncHandler(async (req, res) => {
   if(!goal){
     res.status(400)
     throw new Error('Goal not found!')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  if(!user){
+    res.status(401)
+    throw new Error('User not found!')
+  }
+
+  if(goal.user.toString() != user.id){
+    res.status(401)
+    throw new Error('User not authorized!')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, {
@@ -47,7 +60,7 @@ const updateGoal = asyncHandler(async (req, res) => {
 })
 
 // @desc    Delete goals
-// @route   DELETE /api/goal/:id
+// @route   DELETE /api/goals/:id
 // @access  Private
 const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id)
@@ -55,6 +68,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if(!goal){
     res.status(400)
     throw new Error('Goal not found!')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  if(!user){
+    res.status(401)
+    throw new Error('User not found!')
+  }
+
+  if(goal.user.toString() != user.id){
+    res.status(401)
+    throw new Error('User not authorized!')
   }
 
   const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
